fix: 10 bug fixes — tab mode switch, spinner crash, dead code, mention stripping, diff overflow, provider config, cheerio import, duplicate loader case, unreachable api.js code, unused variable

+++ src/diff.js
@@ -36,7 +36,7 @@ export function unifiedDiff(oldContent, newContent, opts = {}) {
 
 function buildLCS(a, b) {
   const m = a.length, n = b.length;
-  const dp = Array.from({ length: m + 1 }, () => new Uint16Array(n + 1));
+  const dp = Array.from({ length: m + 1 }, () => new Uint32Array(n + 1));
   for (let i = 1; i <= m; i++) {
     for (let j = 1; j <= n; j++) {
       dp[i][j] = a[i - 1] === b[j - 1] ? dp[i - 1][j - 1] + 1 : Math.max(dp[i - 1][j], dp[i][j - 1]);
